Strip debuginfo and enable thinlto in release mode
Reduces the size of the resulting binary from 4.6 MiB to 1.3 MiB at the cost of 200-400 ms of
compile time.
Using fat lto would save an additional 220 KiB, but at the cost of ~7s or ~46% longer compile time.
Stripping symbols too could save another 165 KiB and bring the binary down to 958 KiB, but breaking
cargo-bloat on cargo-bloat isn't worth it.